backend/requirements.txt
fastapi
uvicorn
pypdf2
python-multipart
python-dotenv
groq